chore: add model and view folders
