+++ FitnessApp/Home/Models/Activity.swift
@@ -1,0 +1,17 @@
+//
+//  Activity.swift
+//  FitnessApp
+//
+//  Created by Abdul Suleiman on 04/08/25.
+//
+
+import SwiftUI
+
+struct Activity{
+    let id: Int
+    let title: String
+    let subtitle: String
+    let image: String
+    let tintColor: Color
+    let amount: String
+}
