+++ FitnessApp/Home/Models/Workout.swift
@@ -1,0 +1,19 @@
+//
+//  Workout.swift
+//  FitnessApp
+//
+//  Created by Abdul Suleiman on 04/08/25.
+//
+
+import SwiftUI
+
+struct Workout {
+    let id: Int
+    let title: String
+    let subtitle: String
+    let image: String
+    let date: String
+    let calories: String
+    let tintColor: Color
+
+}
